requirements.txt
absl-py==1.2.0
addict==2.4.0
albumentations==0.5.2
backcall==0.2.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cycler==0.11.0
decorator==5.1.1
docker-pycreds==0.4.0
fonttools==4.33.3
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.11.1
google-auth-oauthlib==0.4.6
grpcio==1.49.0
idna==3.3
imageio==2.19.2
imgaug==0.4.0
importlib-metadata==4.12.0
ipython==7.34.0
jedi==0.18.1
joblib==1.1.0
kiwisolver==1.4.2
lietorch==0.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.4.3
matplotlib-inline==0.1.6
mmcv==1.6.1
networkx==2.6.3
numpy==1.16.1
oauthlib==3.2.1
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
promise==2.3
prompt-toolkit==3.0.31
protobuf==3.19.5
psutil==5.9.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.7.3
sentry-sdk==1.6.0
setproctitle==1.2.3
Shapely==1.8.4
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
tdqm==0.0.1
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.6.7
torch @ file:///home/Process3/KITTI_depth/venv_2/torch-1.8.1%2Bcu111-cp37-cp37m-linux_x86_64.whl
torchvision @ file:///home/Process3/KITTI_depth/venv_2/torchvision-0.9.1%2Bcu111-cp37-cp37m-linux_x86_64.whl
tqdm==4.64.0
traitlets==5.4.0
typing-extensions==4.2.0
urllib3==1.26.10
wandb==0.12.21
wcwidth==0.2.5
Werkzeug==2.2.2
yacs==0.1.8
yapf==0.32.0
zipp==3.8.0
